+++ mymixtape-api/controllers/playlist.go
@@ -1,4 +1,4 @@
-package server
+package controllers
 
 import (
 	"net/http"
