+++ mymixtape-api/controllers/user.go
@@ -1,4 +1,4 @@
-package server
+package controllers
 
 import (
 	"net/http"
