--- mymixtape-api/main.go
@@ -1,12 +1,0 @@
-package main
-
-func main() {
-	// Set Config
-	InitConfig()
-
-	// Setup Router
-	SetupRouter()
-
-	// Run Server
-	StartRouter()
-}
