+++ mymixtape-api/models/spotify.go
@@ -1,0 +1,1 @@
+package models
