+++ mymixtape-api/server.go
@@ -1,0 +1,11 @@
+package main
+
+import "github.com/mymixtape-api/router"
+
+func main() {
+	
+	engine := router.InitRoutes()
+
+	panic(engine.Run())
+
+}
